исправляет ошибки в ветке master
